Refactor user's 2FA settings into SettingsController
